+++ dbx/api/context.py
@@ -2,7 +2,7 @@ import json
 import pathlib
 import time
 from base64 import b64encode
-from pathlib import Path
+from pathlib import Path, PurePosixPath
 from typing import Optional, List, Any
 
 from databricks_cli.sdk import ApiClient
@@ -119,8 +119,8 @@ class RichExecutionContextClient:
     def __init__(self, v2_client: ApiClient, cluster_id: str, language: str = "python"):
         self._client = LowLevelExecutionContextClient(v2_client, cluster_id, language)
 
-    def install_package(self, package_file: Path):
-        installation_command = f"%pip install --force-reinstall {package_file.absolute()}"
+    def install_package(self, package_file: str):
+        installation_command = f"%pip install --force-reinstall {package_file}"
         self._client.execute_command(installation_command, verbose=False)
 
     def setup_arguments(self, arguments: List[Any]):
@@ -147,12 +147,12 @@ class RichExecutionContextClient:
     def client(self):
         return self._client
 
-    def get_temp_dir(self) -> Path:
+    def get_temp_dir(self) -> str:
         command = """
         from tempfile import mkdtemp
         print(mkdtemp())
         """
-        return Path(self._client.execute_command(command, verbose=False))
+        return self._client.execute_command(command, verbose=False)
 
     def remove_dir(self, _dir: str):
         command = f"""
@@ -161,7 +161,7 @@ class RichExecutionContextClient:
         """
         self._client.execute_command(command, verbose=False)
 
-    def upload_file(self, file: Path, prefix_dir: Path) -> Path:
+    def upload_file(self, file: str, prefix_dir: str) -> Path:
         _contents = file.read_bytes()
         contents = b64encode(_contents)
         command = f"""